Split glTF validation directories
