+++ src/validation/gltf/GltfData.ts
