+++ src/validation/gltf/ImageData.ts
