+++ src/validation/gltf/ImageDataReader.ts
