+++ src/validation/gltf/TextureValidator.ts
