+++ src/validation/gltf/structuralMetadata/PropertyTextureValidator.ts
@@ -1,44 +1,49 @@
 import { defined } from "3d-tiles-tools";
 import { defaultValue } from "3d-tiles-tools";
+import { MetadataUtilities } from "3d-tiles-tools";
 import { Schema } from "3d-tiles-tools";
 
-import { ValidationContext } from "../ValidationContext";
-import { BasicValidator } from "../BasicValidator";
-import { RootPropertyValidator } from "../RootPropertyValidator";
-import { ExtendedObjectsValidators } from "../ExtendedObjectsValidators";
+import { ValidationContext } from "../../ValidationContext";
+import { BasicValidator } from "../../BasicValidator";
+import { RootPropertyValidator } from "../../RootPropertyValidator";
+import { ExtendedObjectsValidators } from "../../ExtendedObjectsValidators";
 
-import { MetadataStructureValidator } from "../metadata/MetadataStructureValidator";
+import { MetadataStructureValidator } from "../../metadata/MetadataStructureValidator";
 
-import { PropertyAttributePropertyValidator } from "./PropertyAttributePropertyValidator";
+import { PropertyTexturePropertyValidator } from "./PropertyTexturePropertyValidator";
 
 /**
- * A class for validations related to `propertyAttribute` objects.
+ * A class for validations related to `propertyTexture` objects.
  *
  * This class performs the basic JSON-level validation of the
- * property attributes.
+ * property texture.
  *
  * The validation of any of the underlying binary data of
- * a property attribute has to start at the mesh primitive
- * that refers to the property attribute, because it requires
- * knowledge about the attributes that are defined in the
- * referring mesh primitive.
+ * a property texture has to start at the mesh primitive
+ * that refers to the property texture, because it requires
+ * knowledge about the attributes (texture coordinates)
+ * that are defined in the referring mesh primitive, and
+ * the glTF texture that the definition refers to.
  *
  * @internal
  */
-export class PropertyAttributeValidator {
+export class PropertyTextureValidator {
   /**
    * Performs the validation to ensure that the given object is a
-   * valid `propertyAttribute` object.
+   * valid `propertyTexture` object.
    *
    * @param path - The path for the `ValidationIssue` instances
-   * @param propertyAttribute - The object to validate
+   * @param propertyTexture - The object to validate
+   * @param gltf - The glTF object that contains the definitions
+   * @param meshPrimitive - The mesh primitive that contains the extension
    * @param schema - The `Schema` object
    * @param context - The `ValidationContext` that any issues will be added to
    * @returns Whether the object was valid
    */
-  static validatePropertyAttribute(
+  static validatePropertyTexture(
     path: string,
-    propertyAttribute: any,
+    propertyTexture: any,
+    gltf: any,
     schema: Schema,
     context: ValidationContext
   ): boolean {
@@ -46,8 +51,8 @@ export class PropertyAttributeValidator {
     if (
       !BasicValidator.validateObject(
         path,
-        "propertyAttribute",
-        propertyAttribute,
+        "propertyTexture",
+        propertyTexture,
         context
       )
     ) {
@@ -60,8 +65,8 @@ export class PropertyAttributeValidator {
     if (
       !RootPropertyValidator.validateRootProperty(
         path,
-        "propertyAttribute",
-        propertyAttribute,
+        "propertyTexture",
+        propertyTexture,
         context
       )
     ) {
@@ -73,7 +78,7 @@ export class PropertyAttributeValidator {
     if (
       !ExtendedObjectsValidators.validateExtendedObject(
         path,
-        propertyAttribute,
+        propertyTexture,
         context
       )
     ) {
@@ -81,20 +86,20 @@ export class PropertyAttributeValidator {
     }
     // If there was an extension validator that overrides the
     // default validation, then skip the remaining validation.
-    if (ExtendedObjectsValidators.hasOverride(propertyAttribute)) {
+    if (ExtendedObjectsValidators.hasOverride(propertyTexture)) {
       return result;
     }
 
     // Validate that the class and properties are structurally
     // valid and comply to the metadata schema
-    const className = propertyAttribute.class;
-    const attributeProperties = propertyAttribute.properties;
+    const className = propertyTexture.class;
+    const textureProperties = propertyTexture.properties;
     if (
       !MetadataStructureValidator.validateMetadataStructure(
         path,
-        "property attribute",
+        "property texture",
         className,
-        attributeProperties,
+        textureProperties,
         schema,
         context
       )
@@ -108,7 +113,7 @@ export class PropertyAttributeValidator {
     if (
       !BasicValidator.validateOptionalString(
         path,
-        propertyAttribute,
+        propertyTexture,
         "name",
         context
       )
@@ -119,7 +124,7 @@ export class PropertyAttributeValidator {
     // Here, the basic structure of the class and properties
     // have been determined to be valid. Continue to validate
     // the values of the properties.
-    const validProperties = defaultValue(attributeProperties, {});
+    const validProperties = defaultValue(textureProperties, {});
     const validPropertyNames = Object.keys(validProperties);
     const classes = defaultValue(schema.classes, {});
     const metadataClass = classes[className];
@@ -129,14 +134,10 @@ export class PropertyAttributeValidator {
     for (const propertyName of validPropertyNames) {
       const propertyPath = path + "/properties/" + propertyName;
       const classProperty = classProperties[propertyName];
-
-      let enumValueType = undefined;
-      const enumType = classProperty.enumType;
-      if (defined(enumType)) {
-        const enums = defaultValue(schema.enums, {});
-        const metadataEnum = enums[enumType];
-        enumValueType = metadataEnum.valueType || "UINT16";
-      }
+      const enumValueType = MetadataUtilities.computeEnumValueType(
+        schema,
+        classProperty
+      );
 
       // Note: The check whether 'required' properties are
       // present and have values was already done by the
@@ -144,10 +145,11 @@ export class PropertyAttributeValidator {
       const propertyValue = validProperties[propertyName];
       if (defined(propertyValue)) {
         if (
-          !PropertyAttributePropertyValidator.validatePropertyAttributeProperty(
+          !PropertyTexturePropertyValidator.validatePropertyTextureProperty(
             propertyPath,
             propertyName,
             propertyValue,
+            gltf,
             classProperty,
             enumValueType,
             context
